fix: get balance"
