+++ hooks/useLoadToken.ts
@@ -1,24 +1,38 @@
-'use client';
+"use client";
 
-import { cosmosTokens, oraichainTokens, tokenMap } from '@/constants/bridgeTokens';
-import { chainInfos } from '@/constants/chainInfo';
-import { CW20_TON_CONTRACT, TonNetwork, TonTokensContract, network } from '@/constants/networks';
-import { TonTokenList } from '@/constants/tokens';
-import { genAddressCosmos, handleCheckWallet } from '@/helper';
-import { useAmountsCache, useTokenActions } from '@/stores/token/selector';
-import { fromBinary, toBinary } from '@cosmjs/cosmwasm-stargate';
-import { StargateClient } from '@cosmjs/stargate';
-import { MulticallQueryClient } from '@oraichain/common-contracts-sdk';
-import { OraiswapTokenTypes } from '@oraichain/oraidex-contracts-sdk';
-import { useEffect } from 'react';
+import {
+  cosmosTokens,
+  oraichainTokens,
+  tokenMap,
+} from "@/constants/bridgeTokens";
+import { chainInfos } from "@/constants/chainInfo";
+import {
+  CW20_TON_CONTRACT,
+  TON_ADDRESS_CONTRACT,
+  TonNetwork,
+  TonTokensContract,
+  network,
+} from "@/constants/networks";
+import { TonTokenList } from "@/constants/tokens";
+import { genAddressCosmos, handleCheckWallet } from "@/helper";
+import { useAmountsCache, useTokenActions } from "@/stores/token/selector";
+import { fromBinary, toBinary } from "@cosmjs/cosmwasm-stargate";
+import { StargateClient } from "@cosmjs/stargate";
+import { MulticallQueryClient } from "@oraichain/common-contracts-sdk";
+import { OraiswapTokenTypes } from "@oraichain/oraidex-contracts-sdk";
+import { useEffect } from "react";
 
-import { CW20_DECIMALS, toDisplay } from '@oraichain/oraidex-common';
-import { JettonMinter, JettonWallet } from '@oraichain/ton-bridge-contracts';
-import { getHttpEndpoint } from '@orbs-network/ton-access';
-import { Address } from '@ton/core';
-import { TonClient } from '@ton/ton';
+import { CW20_DECIMALS, toDisplay } from "@oraichain/oraidex-common";
+import { JettonMinter, JettonWallet } from "@oraichain/ton-bridge-contracts";
+import { getHttpEndpoint } from "@orbs-network/ton-access";
+import { Address } from "@ton/core";
+import { TonClient } from "@ton/ton";
 
-async function loadNativeBalance(dispatch: (amount: AmountDetails) => void, address: string, tokenInfo: { chainId: string; rpc: string }) {
+async function loadNativeBalance(
+  dispatch: (amount: AmountDetails) => void,
+  address: string,
+  tokenInfo: { chainId: string; rpc: string }
+) {
   if (!address) return;
   try {
     const client = await StargateClient.connect(tokenInfo.rpc);
@@ -30,38 +44,52 @@ async function loadNativeBalance(dispatch: (amount: AmountDetails) => void, addr
     cosmosTokens
       .filter((t) => t.chainId === tokenInfo.chainId && !t.contractAddress)
       .forEach((t) => {
-        amountDetails[t.denom] = '0';
+        amountDetails[t.denom] = "0";
       });
 
-    const tokensAmount = amountAll.filter((coin) => tokenMap[coin.denom]).map((coin) => [coin.denom, coin.amount]);
+    const tokensAmount = amountAll
+      .filter((coin) => tokenMap[coin.denom])
+      .map((coin) => [coin.denom, coin.amount]);
     Object.assign(amountDetails, Object.fromEntries(tokensAmount));
 
     dispatch(amountDetails);
   } catch (ex) {
-    console.trace('errror');
+    console.trace("errror");
     console.log(ex);
   }
 }
 
 const timer = {};
 
-async function loadTokensCosmos(dispatch: (amount: AmountDetails) => void, kwtAddress: string, oraiAddress: string) {
+async function loadTokensCosmos(
+  dispatch: (amount: AmountDetails) => void,
+  kwtAddress: string,
+  oraiAddress: string
+) {
   if (!kwtAddress && !oraiAddress) return;
   await handleCheckWallet();
   const cosmosInfos = chainInfos.filter(
     (chainInfo) =>
-      (chainInfo.networkType === 'cosmos' || chainInfo.bip44.coinType === 118) &&
+      (chainInfo.networkType === "cosmos" ||
+        chainInfo.bip44.coinType === 118) &&
       // TODO: ignore oraibtc
-      chainInfo.chainId !== ('oraibtc-mainnet-1' as string)
+      chainInfo.chainId !== ("oraibtc-mainnet-1" as string)
   );
   for (const chainInfo of cosmosInfos) {
-    const { cosmosAddress } = genAddressCosmos(chainInfo, kwtAddress, oraiAddress);
+    const { cosmosAddress } = genAddressCosmos(
+      chainInfo,
+      kwtAddress,
+      oraiAddress
+    );
     if (!cosmosAddress) continue;
     loadNativeBalance(dispatch, cosmosAddress, chainInfo);
   }
 }
 
-async function loadCw20Balance(dispatch: (amount: AmountDetails) => void, address: string) {
+async function loadCw20Balance(
+  dispatch: (amount: AmountDetails) => void,
+  address: string
+) {
   if (!address) return;
 
   // get all cw20 token contract
@@ -69,18 +97,18 @@ async function loadCw20Balance(dispatch: (amount: AmountDetails) => void, addres
     ...[
       ...oraichainTokens,
       {
-        name: 'Ton',
-        symbol: 'TON',
+        name: "Ton",
+        symbol: "TON",
         contractAddress: CW20_TON_CONTRACT,
-        denom: 'cw20_ton',
-        coingeckoId: 'the-open-network',
-        decimal: 9
-      }
-    ].filter((t) => t.contractAddress)
+        denom: "cw20_ton",
+        coingeckoId: "the-open-network",
+        decimal: 9,
+      },
+    ].filter((t) => t.contractAddress),
   ];
 
   const data = toBinary({
-    balance: { address }
+    balance: { address },
   });
 
   const multicall = new MulticallQueryClient(window.client, network.multicall);
@@ -88,8 +116,8 @@ async function loadCw20Balance(dispatch: (amount: AmountDetails) => void, addres
   const res = await multicall.aggregate({
     queries: cw20Tokens.map((t) => ({
       address: t.contractAddress,
-      data
-    }))
+      data,
+    })),
   });
 
   const amountDetails = Object.fromEntries(
@@ -97,7 +125,9 @@ async function loadCw20Balance(dispatch: (amount: AmountDetails) => void, addres
       if (!res.return_data[ind].success) {
         return [t.denom, 0];
       }
-      const balanceRes = fromBinary(res.return_data[ind].data) as OraiswapTokenTypes.BalanceResponse;
+      const balanceRes = fromBinary(
+        res.return_data[ind].data
+      ) as OraiswapTokenTypes.BalanceResponse;
       const amount = balanceRes.balance;
       return [t.denom, amount];
     })
@@ -108,7 +138,11 @@ async function loadCw20Balance(dispatch: (amount: AmountDetails) => void, addres
   return amountDetails;
 }
 
-async function loadCw20BalanceWithSpecificTokens(dispatch: (amount: AmountDetails) => void, address: string, specificTokens: string[]) {
+async function loadCw20BalanceWithSpecificTokens(
+  dispatch: (amount: AmountDetails) => void,
+  address: string,
+  specificTokens: string[]
+) {
   if (!address) return;
 
   // get all cw20 token contract
@@ -116,18 +150,20 @@ async function loadCw20BalanceWithSpecificTokens(dispatch: (amount: AmountDetail
     ...[
       ...oraichainTokens,
       {
-        name: 'Ton',
-        symbol: 'TON',
+        name: "Ton",
+        symbol: "TON",
         contractAddress: CW20_TON_CONTRACT,
-        denom: 'cw20_ton',
-        coingeckoId: 'the-open-network',
-        decimal: 9
-      }
-    ].filter((t) => t.contractAddress && specificTokens.includes(t.contractAddress))
+        denom: "cw20_ton",
+        coingeckoId: "the-open-network",
+        decimal: 9,
+      },
+    ].filter(
+      (t) => t.contractAddress && specificTokens.includes(t.contractAddress)
+    ),
   ];
 
   const data = toBinary({
-    balance: { address }
+    balance: { address },
   });
 
   const multicall = new MulticallQueryClient(window.client, network.multicall);
@@ -135,8 +171,8 @@ async function loadCw20BalanceWithSpecificTokens(dispatch: (amount: AmountDetail
   const res = await multicall.aggregate({
     queries: cw20Tokens.map((t) => ({
       address: t.contractAddress,
-      data
-    }))
+      data,
+    })),
   });
 
   const amountDetails = Object.fromEntries(
@@ -144,7 +180,9 @@ async function loadCw20BalanceWithSpecificTokens(dispatch: (amount: AmountDetail
       if (!res.return_data[ind].success) {
         return [t.denom, 0];
       }
-      const balanceRes = fromBinary(res.return_data[ind].data) as OraiswapTokenTypes.BalanceResponse;
+      const balanceRes = fromBinary(
+        res.return_data[ind].data
+      ) as OraiswapTokenTypes.BalanceResponse;
       const amount = balanceRes.balance;
       return [t.denom, amount];
     })
@@ -170,7 +208,7 @@ async function loadCw20BalanceWithSpecificTokens(dispatch: (amount: AmountDetail
 
 export const useLoadTonBalance = ({
   tonAddress,
-  tonNetwork = TonNetwork.Mainnet
+  tonNetwork = TonNetwork.Mainnet,
 }: {
   tonAddress: string;
   tonNetwork?: TonNetwork;
@@ -183,39 +221,45 @@ export const useLoadTonBalance = ({
       // get the decentralized RPC endpoint
       const endpoint = await getHttpEndpoint();
       const client = new TonClient({
-        endpoint
+        endpoint,
       });
-      if (address) {
+      if (address === TON_ADDRESS_CONTRACT) {
         const balance = await client.getBalance(Address.parse(tonAddress));
 
         handleSetTonAmountsCache({
-          ['native_ton']: toDisplay(balance || '0').toString()
+          ["native_ton"]: toDisplay(balance || "0").toString(),
         });
         return {
           balance: balance,
-          jettonWalletAddress: ''
+          jettonWalletAddress: "",
         };
       }
 
-      const token = TonTokenList(tonNetwork).find((e) => e.contractAddress === address);
+      const token = TonTokenList(tonNetwork).find(
+        (e) => e.contractAddress === address
+      );
 
-      const jettonMinter = JettonMinter.createFromAddress(Address.parse(address));
+      const jettonMinter = JettonMinter.createFromAddress(
+        Address.parse(address)
+      );
       const jettonMinterContract = client.open(jettonMinter);
-      const jettonWalletAddress = await jettonMinterContract.getWalletAddress(Address.parse(tonAddress));
+      const jettonWalletAddress = await jettonMinterContract.getWalletAddress(
+        Address.parse(tonAddress)
+      );
       const jettonWallet = JettonWallet.createFromAddress(jettonWalletAddress);
       const jettonWalletContract = client.open(jettonWallet);
       const balance = await jettonWalletContract.getBalance();
 
       handleSetTonAmountsCache({
-        [token.denom]: toDisplay(balance.amount || '0').toString()
+        [token.denom]: toDisplay(balance.amount || "0").toString(),
       });
 
       return {
         balance: balance.amount,
-        jettonWalletAddress
+        jettonWalletAddress,
       };
     } catch (error) {
-      console.log('error load ton balance', error);
+      console.log("error load ton balance", error);
       return {};
     }
   };
@@ -223,41 +267,52 @@ export const useLoadTonBalance = ({
   const loadAllBalanceTonToken = async () => {
     // if (!address) return;
     const allTokens = Object.values(TonTokensContract[tonNetwork]);
+    const endpoint = await getHttpEndpoint();
+    const client = new TonClient({
+      endpoint,
+    });
 
     const fullData = await Promise.all(
       allTokens.map(async (item) => {
-        const endpoint = await getHttpEndpoint();
-        const client = new TonClient({
-          endpoint
-        });
-
-        if (!item) {
+        if (item === TON_ADDRESS_CONTRACT) {
+          // native token: TON
           const balance = await client.getBalance(Address.parse(tonAddress));
 
           return {
             balance: balance,
-            jettonWalletAddress: '',
-            token: item
+            jettonWalletAddress: "",
+            token: item,
           };
         }
-        const jettonMinter = JettonMinter.createFromAddress(Address.parse(item));
+        const jettonMinter = JettonMinter.createFromAddress(
+          Address.parse(item)
+        );
+
         const jettonMinterContract = client.open(jettonMinter);
-        const jettonWalletAddress = await jettonMinterContract.getWalletAddress(Address.parse(tonAddress));
-        const jettonWallet = JettonWallet.createFromAddress(jettonWalletAddress);
+
+        const jettonWalletAddress = await jettonMinterContract.getWalletAddress(
+          Address.parse(tonAddress)
+        );
+
+        // console.log("294-jettonWalletAddress", jettonWalletAddress);
+        const jettonWallet =
+          JettonWallet.createFromAddress(jettonWalletAddress);
         const jettonWalletContract = client.open(jettonWallet);
         const balance = await jettonWalletContract.getBalance();
 
         return {
           balance: balance.amount,
           jettonWalletAddress,
-          token: item
+          token: item,
         };
       })
     );
 
     let amountDetail: AmountDetails = {};
     fullData?.map((data) => {
-      const token = TonTokenList(tonNetwork).find((e) => e.contractAddress === data.token);
+      const token = TonTokenList(tonNetwork).find(
+        (e) => e.contractAddress === data.token
+      );
 
       amountDetail = {
         ...amountDetail,
@@ -265,7 +320,7 @@ export const useLoadTonBalance = ({
         //   data.balance || "0",
         //   token.decimal || CW20_DECIMALS
         // ).toString(),
-        [token.denom]: (data.balance || '0').toString()
+        [token.denom]: (data.balance || "0").toString(),
       };
     });
 
@@ -274,16 +329,26 @@ export const useLoadTonBalance = ({
 
   // @dev: this function will changed based on token minter address (which is USDT, USDC, bla bla bla)
   useEffect(() => {
-    if (tonAddress) loadAllBalanceTonToken();
+    try {
+      if (tonAddress) {
+        loadAllBalanceTonToken();
+      }
+    } catch (error) {
+      console.log("error :>> loadAllBalanceTonToken", error);
+    }
   }, [tonAddress, tonNetwork]);
 
   return {
     loadBalanceByToken,
-    loadAllBalanceTonToken
+    loadAllBalanceTonToken,
   };
 };
 
-const loadTonBalance = (dispatch: (amount: AmountDetails) => void, address: string, tonNetwork: TonNetwork = TonNetwork.Mainnet) => {
+const loadTonBalance = (
+  dispatch: (amount: AmountDetails) => void,
+  address: string,
+  tonNetwork: TonNetwork = TonNetwork.Mainnet
+) => {
   return {};
 };
 
@@ -294,14 +359,18 @@ export const useLoadToken = () => {
   //   const tonAddress = useAuthTonAddress();
 
   const loadToken = ({
-    oraiAddress
+    oraiAddress,
   }: // tonAddress,
   {
     oraiAddress?: string;
     // tonAddress?: string;
   }) => {
     if (oraiAddress) {
-      loadNativeBalance((amounts) => handleSetAmountsCache(amounts), oraiAddress, { chainId: network.chainId, rpc: network.rpc });
+      loadNativeBalance(
+        (amounts) => handleSetAmountsCache(amounts),
+        oraiAddress,
+        { chainId: network.chainId, rpc: network.rpc }
+      );
       loadCw20Balance((amounts) => handleSetAmountsCache(amounts), oraiAddress);
     }
 
@@ -314,6 +383,6 @@ export const useLoadToken = () => {
   };
 
   return {
-    loadToken
+    loadToken,
   };
 };
